refactor(fetch_azure): trim public API and split tests
Address PR review feedback:
- drop AzureHttpClient::inner and into_inner (not needed publicly)
- drop the new_async_runtime free function; add From<Runtime> for
  Arc<dyn AsyncRuntime> to mirror the client's From-based Arc conversion
- split integration tests into tests/client.rs and tests/runtime.rs

client.rs and runtime.rs remain at 100% coverage.
